pyproject.toml
# This section defines the build system using Hatchling
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "basalt_sdk"
dynamic = ["version"]
description = "Basalt SDK for python"
readme = "README.md" # Replaces long_description
license = "MIT"
keywords = ["basalt", "ai", "sdk", "python"]
authors = [
    { name = "Basalt", email = "support@getbasalt.ai" },
]
requires-python = ">=3.10"
classifiers = [
    "Programming Language :: Python :: 3",
    "Operating System :: OS Independent",
]

dependencies = [
    "jinja2>=3.1.6",
    "opentelemetry-instrumentation~=0.59b0",
    "opentelemetry-instrumentation-httpx~=0.59b0",
    "opentelemetry-semantic-conventions~=0.59b0",
    "opentelemetry-api~=1.39.1",
    "opentelemetry-sdk~=1.39.1",
    "opentelemetry-exporter-otlp~=1.39.1",
    "wrapt~=1.17.3",
    "httpx>=0.28.1",
    "black>=26.1.0",
    "typing_extensions>=4.10.0",
]

# This section lists URLs, replacing the `url` parameter
[project.urls]
Homepage = "https://github.com/basalt-ai/basalt-python"

# This replaces dev-requirements.txt
[project.optional-dependencies]
# LLM Provider Instrumentations
openai = [
    "opentelemetry-instrumentation-openai~=0.51.0",
]
anthropic = [
    "opentelemetry-instrumentation-anthropic~=0.51.0",
]
google-generativeai = [
    "opentelemetry-instrumentation-google-generativeai~=0.51.0",
]
# Note: google-genai instrumentation not yet available in openllmetry
# google-genai = [
#     "opentelemetry-instrumentation-google-genai~=0.5b0",
# ]

bedrock = [
    "opentelemetry-instrumentation-bedrock~=0.51.0",
]
vertex-ai = [
    "opentelemetry-instrumentation-vertexai~=0.51.0",
]

mistralai = [
    "opentelemetry-instrumentation-mistralai~=0.51.0",
]


# Vector Database Instrumentations
chromadb = [
    "opentelemetry-instrumentation-chromadb~=0.51.0",
]
pinecone = [
    "opentelemetry-instrumentation-pinecone~=0.51.0",
]
qdrant = [
    "opentelemetry-instrumentation-qdrant~=0.51.0",
]

# Framework Instrumentations
langchain = [
    "opentelemetry-instrumentation-langchain~=0.51.0",
]
llamaindex = [
    "opentelemetry-instrumentation-llamaindex~=0.51.0",
]

# Mising Framework Instrumentation
# langgraph, litellm


# Convenience groups
llm-all = [
    "basalt_sdk[openai,anthropic,google-generativeai,bedrock,vertex-ai,mistralai]",
]
vector-all = [
    "basalt_sdk[chromadb,pinecone,qdrant]",
]
framework-all = [
    "basalt_sdk[langchain,llamaindex]",
]
all = [
    "basalt_sdk[llm-all,vector-all,framework-all]",
]

# Development dependencies
dev = [
    # Build tools
    "setuptools",
    "wheel",
    "twine",
    # Test dependencies
    "parameterized",
    "requests", # Often included for testing API clients
    # Modern tooling
    "ruff",
    "pytest",
    "pytest-cov",
    "pytest-asyncio",
    "coverage",
    "mistralai",

    "google-genai",
    "anthropic",
    "openai",
    "opentelemetry-instrumentation-anthropic",
    "opentelemetry-instrumentation-google-generativeai",
    "opentelemetry-instrumentation-google-genai",
    "opentelemetry-instrumentation-openai",
    "opentelemetry-instrumentation-vertexai",
]


# A recommended configuration for Ruff, the modern linter/formatter
[tool.ruff]
line-length = 160
target-version = "py312"
# Enable lint + format in one tool
lint.select = ["E", "F", "I", "B", "UP", "ANN"]
lint.ignore = ["ANN101", "ANN102"]  # example: ignore self/cls annotations
lint.fixable = ["ALL"]
lint.unfixable = []

[tool.ruff.lint.isort]
known-first-party = ["basalt"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
line-ending = "lf"
docstring-code-format = true

# Hatch configuration
[tool.hatch.version]
path = "basalt/_version.py"

[tool.hatch.build.targets.wheel]
packages = ["basalt"]

[tool.hatch.build.targets.sdist]
include = [
    "/basalt",
    "/tests",
    "/docs",
    "/README.md",
    "/pyproject.toml",
]

# Hatch environments for development and testing
[tool.hatch.envs.default]
dependencies = [
    "pytest>=9.0.2",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "anyio>=4.12.1",
    "ruff>=0.14.13",
    "mypy>=1.8.0",
]

[tool.hatch.envs.default.scripts]
test = "pytest tests/ --cov=basalt --cov-report=term --cov-report=xml"
test-verbose = "pytest tests/ -v --cov=basalt --cov-report=term --cov-report=xml"
lint = "ruff check ."
lint-fix = "ruff check --fix ."
fmt = "ruff format ."
fmt-check = "ruff format --check ."
typecheck = "mypy basalt tests"
all = [
    "fmt",
    "lint-fix",
    "typecheck",
    "test",
]

# Matrix environments for testing across Python versions
[tool.hatch.envs.test]
dependencies = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
    "pytest-asyncio>=0.23.0",
    "anyio>=4.0.0",
]

[[tool.hatch.envs.test.matrix]]
python = ["3.10", "3.11", "3.12", "3.13", "3.14"]

[tool.hatch.envs.test.scripts]
run = "pytest tests/ --cov=basalt --cov-report=term"

# Environment with all optional dependencies for comprehensive testing
[tool.hatch.envs.full]
features = ["all", "dev"]

[tool.hatch.envs.full.scripts]
test = "pytest tests/ --cov=basalt --cov-report=term --cov-report=xml"
